pyproject.toml
[tool.poetry]
name = "civicblocks"
version = "0.1.0"
description = "Open source civic feedback platform for verified Iowa voters"
authors = ["CivicBlocks Contributors"]
license = "MIT"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python          = ">=3.10,<4.0"

# Database
psycopg2-binary = ">=2.9"

# Data pipeline
pandas          = ">=2.2"
geopandas       = ">=0.14"
shapely         = ">=2.0"
pyproj          = ">=3.6"
tqdm            = ">=4.66"

# Networking / APIs
requests        = ">=2.31"
anthropic       = ">=0.28"
voyageai        = ">=0.2"

# Crypto / identity
mnemonic        = ">=0.21"

# API
fastapi         = ">=0.111"
uvicorn         = {extras = ["standard"], version = ">=0.29"}
pyjwt           = ">=2.8"

# Config
python-dotenv   = ">=1.0"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
